SharedWorker: Use CSP headers to outside settings
This CL passes real ContentSecurityPolicy and ContentSecurityPolicyType
values to creation_params which is used for main resource fetching.
It used to be using dummy values.

This changes the behavior.
Web-platform-tests to check its behavior on Module SharedWorker are added.
WPTs for Classic SharedWorker will be added in the following CLs.

Bug: 1049011, 824646
Change-Id: I6fa79b0394426f09b19da1abbdf9ecf483d558ec