[Trusted Types] Only accept parser's HTMLScriptElement::TextChildContent.
A script might get access to an HTMLScriptELement instance before the parser
has finished parsing. In that case, the script might manipulate the content
before HTMLScriptElement::FinishParsingChildren() is called, and we should not
accept the TextChildContent() as having come from the parser. We resolev this
by remembering whether the HTMLScriptElement's children have ever been changed
by the API, and if so don't uppdate the script_text_internal_slot_.

Bug: 1033101
Change-Id: Ia4972b1deb967b82b36b05176cca7d587b73bd50