Reland "Fixes click on inline element embedded in anchor inside <summary> breaks anchor." This reverts commit b4cb14f3cacece63ea28e9a021a77c41867e0503.
Reason for revert: Getting ready to re-land

The original fix for Bug: 538283 introduced a regression, Bug: 1045433.
So it was reverted. This is another fix for the original issue and
 ensures anchor behavior does not regressed.

To recap, the original issue is caused by DOMActivate event generated
from inline elements. DOMActivate bypasses anchor, bubbles up to
<summary>, and gets handled. The underlying event, click, stops
propagating and terminates. The original fix for anchor looks at the
underlying event and handles it if it's click. However, this logic is
wrong because the anchor can preemptively handle a click that
belongs to another element earlier in the event path.

This example, in file label-inside-anchor.html, illustrate the problem.

<a href="javascript:void(0)" target="_blank">
    <label for="peas"><span id="text">peas?</span></label>
    <input type="checkbox" name="peas" id="peas">
</a>

Clicking on 'text' generates DOMActivate, which is handled by <a>.
Prior to my original fix, DOMActivate isn't handled, and click event
bubbles up until it's handle by the <label> attribute. Bug: 1045433 is
another example of the same behavior. Video element wrapped by anchor.

The fix is for anchor element to check DOMActivate event. And if
DOMActivate contains a underlying click, anchor stops DOMActivate's
propagation, and wait for click event and handles that when it arrives.
Therefore, if an element between the target and anchor that should
handle click, it can. However, calling event.stopPropagation() only sets
a flag, the event still bubbles, dom/events/event_dispatcher.cc&l=360.
Calling event('DOMActivate').SetDefaultHandled() will stop the underlying
event too. So the best way is for the summary element to check this flag.

Original change's description:
> Revert "Fixes click on inline element embedded in anchor inside <summary> breaks anchor."
>
> This reverts commit 4eb4df63f17ee0f22cd17472fa6ae92311e49900.
>
> Reason for revert: Caused crbug.com/1045433.
>
> Original change's description:
> > Fixes click on inline element embedded in anchor inside <summary> breaks anchor.
> >
> > Previous to this CL, clicking on an inline element embedded in an anchor placed
> > inside a <summary> tag will expand the <details> section instead of navigating to
> > the anchor's href. However, when the anchor is placed outside of <summary>,
> > it behaves correctly.
> >
> > The error is caused by DOMActivate event generated by the inline element. As
> > DOMActivate bubbles up, it bypasses the anchor's event handler, reaches the
> > <summary>, and is handled there. Once DOMActivate is handled, the original
> > click event stops propagating and terminates. This behavior, however, differs
> > from when the anchor tag is placed outside of the summary. DOMActivate isn't
> > handled, and the original click event keeps bubbling up till it's handled by
> > the anchor.
> >
> > DOMActivate event is deprecated:
> > https://developer.mozilla.org/en-US/docs/Web/API/Element/DOMActivate_event.
> > However, since blink still has code that depends on it, replacing it is outside
> > of the scope for this fix. Instead, this fix is for the anchor element to
> > handle the DOMActivate event as it bubbles up. The anchor event handler checks
> > the underlying event of DOMActivate and handles it if it's a click.
> >
> > I also looked at an alternative fix by trying to prevent the DOMActivate event
> > from bubbling up. But calling event.stopPropagation() doesn't work as the
> > EventDispatcher::DispatchEventPostProcess doesn't check this status.
> >
> >
> > Bug: 538283
> > Change-Id: I11fb072faa0563279d43a28e5dc19cee89906bf0
> > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1928234
> > Reviewed-by: Kent Tamura <tkent@chromium.org>
> > Reviewed-by: Mason Freed <masonfreed@chromium.org>
> > Commit-Queue: Yu Han <yuzhehan@chromium.org>
> > Cr-Commit-Position: refs/heads/master@{#718552}
>>
> Bug: 1045433
> Change-Id: I14b369beb04171ef846cb3a79ebb3fe268cf5c89
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2023267
> Reviewed-by: Mason Freed <masonfreed@chromium.org>
> Commit-Queue: Yu Han <yuzhehan@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#735575}

Bug: 1045433
Change-Id: I0cdd7b48a2c4e9bf2b653edef9ce7970d7e64938